# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoJudge API
  version: 0.1.0
  description: MVP API for problem listing, code submission, and asynchronous judge results.
paths:
  /auth/register:
    post:
      summary: Register a user and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "201":
          description: Registered user
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Username already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/login:
    post:
      summary: Login and create a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthRequest"
      responses:
        "200":
          description: Logged-in user
          headers:
            Set-Cookie:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /auth/logout:
    post:
      summary: Delete current session
      security:
        - cookieAuth: []
      responses:
        "204":
          description: Session deleted
  /auth/me:
    get:
      summary: Get current user
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Login required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /problems/{id}:
    get:
      summary: Get problem detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Problem detail without hidden test cases
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Problem"
        "404":
          description: Problem not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /submissions:
    get:
      summary: List current user's submission history
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Submission history ordered by latest update
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      summary: Create a code submission
      security:
        - cookieAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: Client-generated key that makes concurrent retries return the original submission.
          schema:
            type: string
            minLength: 1
            maxLength: 128
            pattern: "^[A-Za-z0-9][A-Za-z0-9._:-]{0,127}$"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "202":
          description: Submission accepted and queued
          headers:
            Idempotency-Replayed:
              description: Present and true when an existing submission is replayed.
              schema:
                type: boolean
            RateLimit-Limit:
              schema:
                type: integer
            RateLimit-Remaining:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Problem not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Login required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Idempotency key reused with a different request body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Per-user submission rate limit exceeded
          headers:
            Retry-After:
              schema:
                type: integer
            RateLimit-Limit:
              schema:
                type: integer
            RateLimit-Remaining:
              schema:
                type: integer
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Submission admission control is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /submissions/{id}:
    get:
      summary: Get current user's submission result
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Submission status and result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Login required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /submissions/{id}/artifacts:
    get:
      summary: List current user's submission artifact metadata
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Artifact metadata for the submission
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SubmissionArtifact"
        "401":
          description: Login required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Submission not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /submissions/{id}/artifacts/{artifactId}:
    get:
      summary: Download current user's submission artifact content
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: artifactId
          in: path
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: Artifact file content
          content:
            text/plain:
              schema:
                type: string
        "401":
          description: Login required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Submission or artifact not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: Object store is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: gojudge_session
  schemas:
    AuthRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 32
          pattern: "^[A-Za-z0-9_-]{3,32}$"
        password:
          type: string
          minLength: 8
          maxLength: 72
    User:
      type: object
      required:
        - id
        - username
        - createdAt
      properties:
        id:
          type: string
        username:
          type: string
        createdAt:
          type: string
          format: date-time
...
    CreateSubmissionRequest:
      type: object
      required:
        - problemId
        - language
        - code
      properties:
        problemId:
          type: string
          example: sum
        language:
          type: string
          enum:
            - go
            - cpp
            - python
        code:
          type: string
          maxLength: 65536
    Submission:
      type: object
      required:
        - id
        - problemId
        - language
        - status
        - createdAt
        - updatedAt
      properties:
        id:
          type: string
        userId:
          type: string
        problemId:
          type: string
        language:
          type: string
          enum:
            - go
            - cpp
            - python
        status:
          type: string
          enum:
            - queued
            - running
            - accepted
            - wrong_answer
            - compile_error
            - runtime_error
            - time_limit_exceeded
            - internal_error
        result:
          $ref: "#/components/schemas/JudgeResult"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    SubmissionArtifact:
      type: object
      required:
        - id
        - submissionId
        - attempt
        - kind
        - sha256
        - sizeBytes
        - createdAt
      properties:
        id:
          type: integer
          minimum: 1
        submissionId:
          type: string
        attempt:
          type: integer
          minimum: 1
        kind:
          type: string
          enum: [source, stdout, stderr]
        sha256:
          type: string
        sizeBytes:
          type: integer
          minimum: 0
        createdAt:
          type: string
          format: date-time
    Problem:
      type: object
      required:
        - id
        - title
        - description
        - language
        - timeLimitMs
        - memoryLimitMb
        - difficulty
        - collection
        - sortOrder
        - tags
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        language:
          type: string
          enum: [go, cpp, python]
        timeLimitMs:
          type: integer
          minimum: 1
        memoryLimitMb:
          type: integer
          minimum: 1
        difficulty:
          type: string
          enum: [easy, medium, hard]
        collection:
          type: string
          enum: [starter, hot20]
        sortOrder:
          type: integer
          minimum: 0
        tags:
          type: array
          items:
            type: string
    JudgeResult:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          enum: [accepted, wrong_answer, compile_error, runtime_error, time_limit_exceeded, internal_error]
        stdout:
          type: string
        stderr:
          type: string
        exitCode:
          type: integer
        durationMs:
          type: integer
          format: int64
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
            message:
              type: string